Add files via upload
Absolute Java Chapter Two Program 9. I wrote the program so that it would accept "hate" anywhere in the string. As it was previously written, per the instructions of the text, it would not allow for "Hate" in index zero, the beginning of the sentence. So now it recognizes hate in any position and replaces it with "love" with the appropriate capitalization. There may be some discrepancy if the user gets cheeky. Check the program output.